backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
drf-spectacular>=0.27,<1.0
dj-database-url>=2.2,<3.0
gunicorn>=22.0,<24.0
whitenoise>=6.7,<7.0
